AUT-3487: Do not allow deleting a default mfa method
We don't want users to delete default mfa methods as they must always have an mfa method - the flow for removing a primary mfa method will instead be to:

* add a new backup
* switch these methods round by promoting the backup to a default method
* remove the old default method (which will have been relegated to a backup via the step above)

This makes use of the either type that has been introduced to this repo by adding in the vavr library

+++ shared/src/main/java/uk/gov/di/authentication/shared/services/mfa/DynamoMfaMethodsService.java
@@ -74,6 +74,21 @@ public class DynamoMfaMethodsService implements MfaMethodsService {
 
     public Either<MfaDeleteFailureReason, String> deleteMfaMethod(
             String email, String mfaIdentifier) {
+        var mfaMethods = dynamoService.getUserCredentialsFromEmail(email).getMfaMethods();
+
+        var maybeMethodToDelete =
+                mfaMethods.stream()
+                        .filter(mfaMethod -> mfaMethod.getMfaIdentifier().equals(mfaIdentifier))
+                        .findFirst();
+
+        if (maybeMethodToDelete.isPresent()
+                && maybeMethodToDelete
+                        .get()
+                        .getPriority()
+                        .equals(PriorityIdentifier.DEFAULT.name())) {
+            return Either.left(MfaDeleteFailureReason.CANNOT_DELETE_DEFAULT_METHOD);
+        }
+
         dynamoService.deleteMfaMethodByIdentifier(email, mfaIdentifier);
         return Either.right(mfaIdentifier);
     }
